Move connect_services (p2p)
Move connect_services to testing_utils module
