+++ p2p/src/peer_manager/mod.rs
@@ -21,7 +21,6 @@
 
 #![allow(rustdoc::private_intra_doc_links)]
 
-pub mod helpers;
 pub mod peerdb;
 
 use std::{
