+++ p2p/src/testing_utils.rs
@@ -13,6 +13,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#![allow(clippy::unwrap_used)]
+
 use std::{
     net::{IpAddr, Ipv6Addr, SocketAddr},
     time::Duration,
@@ -25,7 +27,7 @@ use crate::net::{
     mock::transport::{
         MockChannelTransport, NoiseEncryptionAdapter, NoiseTcpTransport, TcpTransportSocket,
     },
-    types::ConnectivityEvent,
+    types::{ConnectivityEvent, PeerInfo},
     ConnectivityService, NetworkingService,
 };
 
@@ -61,7 +63,7 @@ impl TestTransportMaker for TestTransportTcp {
     }
 
     fn make_address() -> Self::Address {
-        "[::1]:0".parse().expect("valid address")
+        "[::1]:0".parse().unwrap()
     }
 }
 
@@ -138,6 +140,43 @@ impl RandomAddressMaker for TestChannelAddressMaker {
         let mut rng = make_pseudo_rng();
         rng.gen()
     }
+}
+
+/// Can be used in tests only, will panic in case of errors
+pub async fn connect_services<T>(
+    conn1: &mut T::ConnectivityHandle,
+    conn2: &mut T::ConnectivityHandle,
+) -> (T::Address, PeerInfo<T>, PeerInfo<T>)
+where
+    T: NetworkingService + std::fmt::Debug,
+    T::ConnectivityHandle: ConnectivityService<T>,
+{
+    let addr = timeout(Duration::from_secs(5), conn2.local_addresses())
+        .await
+        .expect("local address fetch not to timeout")
+        .unwrap();
+    conn1.connect(addr[0].clone()).await.expect("dial to succeed");
+
+    let (address, peer_info1) = match timeout(Duration::from_secs(5), conn2.poll_next()).await {
+        Ok(event) => match event.unwrap() {
+            ConnectivityEvent::InboundAccepted { address, peer_info } => (address, peer_info),
+            event => panic!("expected `InboundAccepted`, got {event:?}"),
+        },
+        Err(_err) => panic!("did not receive `InboundAccepted` in time"),
+    };
+
+    let peer_info2 = match timeout(Duration::from_secs(5), conn1.poll_next()).await {
+        Ok(event) => match event.unwrap() {
+            ConnectivityEvent::OutboundAccepted {
+                address: _,
+                peer_info,
+            } => peer_info,
+            event => panic!("expected `OutboundAccepted`, got {event:?}"),
+        },
+        Err(_err) => panic!("did not receive `OutboundAccepted` in time"),
+    };
+
+    (address, peer_info1, peer_info2)
 }
 
 /// Return first event that is accepted by predicate.
