+++ p2p/tests/block_announcement.rs
@@ -21,7 +21,8 @@ use common::{
 };
 
 use p2p::testing_utils::{
-    TestTransportChannel, TestTransportMaker, TestTransportNoise, TestTransportTcp,
+    connect_services, TestTransportChannel, TestTransportMaker, TestTransportNoise,
+    TestTransportTcp,
 };
 use p2p::{
     error::{P2pError, PublishError},
@@ -34,7 +35,6 @@ use p2p::{
         types::SyncingEvent,
         ConnectivityService, NetworkingService, SyncingMessagingService,
     },
-    peer_manager::helpers::connect_services,
 };
 
 // Test announcements with multiple peers and verify that the message validation is done and peers
